feat : 공통 타입, 공통 함수를 types, utils 폴더로 분리
