+++ src/components/Home/ProductCategory/ProductCategoryContainer.tsx
@@ -1,66 +1,9 @@
 import React from "react";
+import { getCategoryList } from "../../../utils/getCategoryList";
 import ProductCategory from "./ProductCategory";
 
-export interface ICategoryList {
-  image: string;
-  text: string;
-  href: string;
-}
-
-const categoryList: ICategoryList[] = [
-  {
-    image: "",
-    text: "전체",
-    href: "",
-  },
-  {
-    image: "",
-    text: "식품",
-    href: "",
-  },
-  {
-    image: "",
-    text: "주방",
-    href: "",
-  },
-  {
-    image: "",
-    text: "욕실",
-    href: "",
-  },
-  {
-    image: "",
-    text: "생활",
-    href: "",
-  },
-  {
-    image: "",
-    text: "취미",
-    href: "",
-  },
-  {
-    image: "",
-    text: "선물",
-    href: "",
-  },
-  {
-    image: "",
-    text: "여성용품",
-    href: "",
-  },
-  {
-    image: "",
-    text: "반려동물",
-    href: "",
-  },
-  {
-    image: "",
-    text: "문구",
-    href: "",
-  },
-];
-
 const ProductCategoryContainer = () => {
+  const categoryList = getCategoryList();
   return <ProductCategory categoryList={categoryList} />;
 };
