+++ src/types/index.ts
@@ -1,0 +1,5 @@
+export interface ICategoryList {
+  image: string;
+  text: string;
+  href: string;
+}
