+++ src/utils/getCategoryList.ts
@@ -1,0 +1,57 @@
+import { ICategoryList } from "../types";
+
+export const getCategoryList = () => {
+  const categoryList: ICategoryList[] = [
+    {
+      image: "",
+      text: "전체",
+      href: "",
+    },
+    {
+      image: "",
+      text: "식품",
+      href: "",
+    },
+    {
+      image: "",
+      text: "주방",
+      href: "",
+    },
+    {
+      image: "",
+      text: "욕실",
+      href: "",
+    },
+    {
+      image: "",
+      text: "생활",
+      href: "",
+    },
+    {
+      image: "",
+      text: "취미",
+      href: "",
+    },
+    {
+      image: "",
+      text: "선물",
+      href: "",
+    },
+    {
+      image: "",
+      text: "여성용품",
+      href: "",
+    },
+    {
+      image: "",
+      text: "반려동물",
+      href: "",
+    },
+    {
+      image: "",
+      text: "문구",
+      href: "",
+    },
+  ];
+  return categoryList;
+};
